arrange code in the proper files
