+++ src/JavaScript-files/home-folder/counter.js
@@ -5,7 +5,7 @@ const itemCounter = async (value) => {
     counter.innerHTML = `Total Pizza: ${value.length}`;
     underText.innerHTML = `Reserve a pizza, come any time inside the reservation and have it in
     20 min after you arrive!`;
-  }, 4000);
+  }, 1200);
 
   return value.length;
 };
