+++ src/JavaScript-files/home-folder/home.js
@@ -1,11 +1,9 @@
 /* eslint-disable import/no-cycle */
-import { getLikes } from './hitApi.js';
+import { clickLikes, getLikes } from '../api-calls/involvementAPI';
 // eslint-disable-next-line import/no-cycle
-import displayPopUp from './display-comments.js';
-import displayResPopUp from './reservation_feature/reserveation.js';
+import displayPopUp from '../comms-folder/display-comments.js';
+import displayResPopUp from '../reservations-folder/reserveation.js';
 import showLiked from './displayLike.js';
-
-const apiEndPoint = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/8FcrK9POw5EbfAJUs4DD/likes';
 
 const displayPizza = document.querySelector('.display-pizza');
 
@@ -48,22 +46,13 @@ const showPizza = async (value) => {
 
     showLiked(likedData, dat, namePizz);
 
-    const clickLikes = async () => {
-      await fetch(apiEndPoint, {
-        method: 'POST',
-        body: JSON.stringify({
-          item_id: dat.id,
-        }),
-        headers: {
-          'Content-Type': 'application/json; charset=utf-8',
-        },
-      });
+    likes.addEventListener('click', (e) => {
+      e.preventDefault();
+      clickLikes(dat);
       setTimeout(() => {
         showLiked(likedData, dat, namePizz);
-      }, 100);
-    };
-
-    likes.addEventListener('click', clickLikes);
+      }, 1000);
+    });
 
     divContainer.insertBefore(namePizz, namePub);
