+++ src/JavaScript-files/home-folder/spinner.js
@@ -3,7 +3,7 @@ const spinLoader = document.querySelector('.loading');
 const loadSpinner = () => {
   setTimeout(() => {
     spinLoader.style.display = 'none';
-  }, 4500);
+  }, 1000);
 };
 
 export default loadSpinner;
