+++ src/__mocks__/itemCounter.js
