+++ src/__mocks__/itemCounter.test.js
