--- src/index.js
@@ -1,7 +1,0 @@
-import './style.css';
-import { getPizza, getLikes } from './hitApi.js';
-import loadSpinner from './spinner.js';
-
-getLikes();
-getPizza();
-loadSpinner();
